+++ src/main/java/net/tundra/core/scene/InterfaceCamera.java
@@ -15,5 +15,5 @@ public final class InterfaceCamera extends Camera {
   }
 
   @Override
-  public void update(Game game, int delta) {}
+  public void update(Game game, float delta) {}
 }
