+++ src/main/java/net/tundra/examples/Floor.java
@@ -16,7 +16,7 @@ public class Floor extends PhysicsObject {
   }
 
   @Override
-  public void update(Game game, int delta) throws TundraException {}
+  public void update(Game game, float delta) throws TundraException {}
 
   @Override
   public void render(Game game, Graphics graphics) throws TundraException {
